WebViewをActivityに変更
