+++ SW_API_Kotlin_Project/app/src/main/AndroidManifest.xml
@@ -23,6 +23,9 @@
             android:name=".activity.planet.PlanetActivity"
             android:exported="false" />
         <activity
+            android:name=".activity.webview.WebViewActivity"
+            android:exported="false" />
+        <activity
             android:name=".activity.MainActivity"
             android:exported="true">
             <intent-filter>
