+++ SW_API_Kotlin_Project/app/src/main/java/com/sw/sw_api_kotlin_project/fragment/others/OthersFragment.kt
@@ -42,7 +42,7 @@ class OthersFragment : BaseFragment() {
             ).show()
         }
         binding.webViewDocumentation.setOnClickListener {
-            val action = OthersFragmentDirections.actionNavOtherToWebView(
+            val action = OthersFragmentDirections.actionNavOtherToNavWebView(
                 WebViewInfo(
                     title = getString(R.string.other_api_document),
                     url = WebViewURL.swApiDocumentation,
